Use creation functions instead of static objects for services
